third_party/WebKit/LayoutTests/imported/wpt/html/semantics/embedded-content/the-object-element/object-attributes-expected.txt

This is a testharness.js-based test.
FAIL object.contentWindow Cannot read property 'name' of undefined
FAIL object.width assert_equals: The width should be 100px. expected "100px" but got "auto"
FAIL object.height assert_equals: The height should be 50px. expected "50px" but got "auto"
Harness: the test ran to completion.
